Add payload extending by zero bytes
In case the agent's data is smaller than MTU value the network checker
will not be able to detect problems with packet fragmentation. Special
parameter is added for CLI which triggers a procedure that finds out MTU value
of the link in communication with the server and extends marshaled
payload by zero bytes slice of that value length.